requirements.txt
annotated-types==0.7.0
anyio==4.10.0
assemblyai==0.42.1
certifi==2025.8.3
decorator==5.2.1
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.2
pillow==11.3.0
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
regex==2025.7.34
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Unidecode==1.4.0
websockets==15.0.1
